Move DisableAiSettings from `client` to `project` crate
This way it is pulled in by Project::init_settings(cx) which is involved in a lot of test initialization
